pyproject.toml
[tool.poetry]
name = "matsuyama"
version = "0.1.0"
description = ""
authors = ["weiwei <wanweiwei07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
matplotlib = "^3.10.0"
pandas = "^2.2.3"
scipy = "^1.14.1"
torch = { version = "*", source = "torch_cu121" }
torchvision = { version = "*", source = "torch_cu121" }
torchaudio = { version = "*", source = "torch_cu121" }
seaborn = "^0.13.2"
japanize-matplotlib = "^1.1.3"
keyboard = "^0.13.5"
nokovpy = {path = "nokovpy-3.0.1-py3-none-any.whl"}
statsmodels = "^0.14.4"
diffusers = ">=0.33.0"
accelerate = "^1.6.0"
datasets = "^3.5.0"
wandb = "^0.19.9"
tensorboard = "^2.19.0"
pyserial = "^3.5"
discord-py = "^2.5.2"
python-dotenv = "^1.1.1"
git-filter-repo = "^2.47.0"

[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
